Create binary.java
This is an program that uses an alternative way to calculate binary, i'm trying to prove by math, using true table. If it work, than we can patent this way to calculate by this different way, even it's more complicated when the numers are bigger, need to add, negative and positives numbers, and numbers with comma.